// frontend/src/types/User.ts
export interface User {
  id: number;
  username: string;
  email: string;
  first_name?: string;
  last_name?: string;
  is_active?: boolean;
  is_verified?: boolean;
  created_at?: string;
  last_login?: string;
  dexcom_username?: string;
  dexcom_ous?: boolean;
  myfitnesspal_username?: string;
  apple_health_authorized?: boolean;
  google_fit_authorized?: boolean;
  fitbit_authorized?: boolean;
  height_cm?: number;
  weight_kg?: number;
  birthdate?: string;
  gender?: string;
  target_glucose_min?: number;
  target_glucose_max?: number;
  insulin_carb_ratio?: number;
  insulin_sensitivity_factor?: number;
  diabetes_type?: number;
  diagnosis_date?: string;
  notification_preferences?: any;
  privacy_preferences?: any;
}